Merged Webpack Configs into webpack.base.js
We used webpack's merge function in order to merge duplicated config settings found in client and server.

We store that into the file webpack.base.js. Inside of there we store the duplicate code and then require webpack.base.js on both client and server.

This allows us to save space and avoid code duplication.