Imported babel-polyfill and successfully render the list of users.
We were able to figure out how to render the list of users using babel-polyfill.

Although we were able to display the list, we apparently are still using client side rendering for this instead of server side rendering which is what we want.

In the next chapter I think this will be addressed.
Section: 6
Lecture: 40