45: We revamped our Routing method (renderRoutes from react-router-config)
In this commit we changed our routing method.

The changes are most clear in the Routes.js file. We switched from  using our route configuration COMPONENT to using an array of route objects instead.

We then had to find every instance that we used our Route component and replace it with a div inside of which is a javascript snippet that calls the renderRoutes function and we pass in our Routes import.

Section: 7
Lecture: 45