53: Page Dir // Refactored Page/Route interaction
Instead of passing in { loadData } separately we now return objects inside of our HomePage/UserListPage and they include the default export along with the loadData.

This is necessary because {loadData} can exist in multiple pages and it could pose a problem if you are importing more than 2 pages that need to load data.

Section: 8
Lecture: 53

EDIT: Amended a missing semi colon that broke my entire program lmfao...