62: Setting up Proxy
In this commit we set up our server proxy using

import proxy from 'express-http-proxy'.

We are grabbing the cookie from the initial client request and attaching it to the request that AXIOS (in our renderer server) is making in order to trick the API into thinking that we are actually the client who made the request instead of AXIOM inside of our render server.

Section 8
Lecture 62