80: Created Admin Feature
Created FETCH_ADMINS action in our action/index.js file. We Created the AdmistListPage Page in  our pages folder. We also created the reducer and properly routed it.

Section 10
Lecture 80

SIDE NOTE: The lesson is a bit dated so the google oath page is not properly working. It cannot authenticate us therefore we cannot check if the login/logout functionality works.

It sucks since that's the whole point of this section but... whatever.